in-process changes for network abstraction
